src/bff-movil/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
python-dotenv==1.1.1
google-cloud-pubsub==2.23.1
debugpy==1.8.16
redis==5.0.1
httpx==0.27.0